Add dedicated AbstractMultiSet serialization test
